// src/lib/types/singleUserType.ts
// Generated by https://quicktype.io

export interface SingleUserType {
  _id: string;
  first_name: string;
  last_name: string;
  accountName: string;
  client_id: string;
  email: string;
  phone: string;
  image: string;
  role: string;
  employment: Employment;
  investment: Investment;
  document: Document;
  status: Status;
  createdAt: string;
  __v: number;
  name: string;
}

export interface Document {
  name: string;
  image: string;
}

export interface Employment {
  annual_income: string;
}

export interface Investment {
  goal: string;
  experience: string;
  marital_status: string;
  next_of_kin_name: string;
  next_of_kin_phone: string;
  next_of_kin_relationship: string;
  next_of_kin_email: string;
}

export interface Status {
  employment: string;
  document: string;
  investment: string;
  bank: string;
  biodata: string;
  access: string;
}